+++ src/hooks/common/useRoute.ts
@@ -12,6 +12,7 @@ export const useRoute = () => {
     navigateToSelectStyle: () => navigate(PATH.SELECT_STYLE),
     navigateToSelectTag: () => navigate(PATH.SELECT_TAG),
     navigateToReviewResult: () => navigate(PATH.REVIEW_RESULT),
+    navigateToLoading: () => navigate(PATH.LOADING),
   };
 
   return routes;
